+++ xrpl/models/base_model.py
@@ -21,7 +21,7 @@ def _key_to_json(field: str) -> str:
     words = split(r"(?=[A-Z])", field)
     lower_words = [word.lower() for word in words if word]
     snaked = "_".join(lower_words)
-    return sub("i_d", "id", snaked)
+    return sub("u_n_l", "unl", sub("i_d", "id", snaked))
 
 
 def _value_to_json(value: str) -> str:
@@ -163,7 +163,7 @@ class BaseModel(ABC):
     def _get_only_init_args(
         cls: Type[BaseModel], args: Dict[str, Any]
     ) -> Dict[str, Any]:
-        init_keys = {field.name for field in fields(cls)}
+        init_keys = {field.name for field in fields(cls) if field.init is True}
         valid_args = {key: value for key, value in args.items() if key in init_keys}
         return valid_args
 
@@ -236,10 +236,12 @@ class BaseModel(ABC):
         Returns:
             The dictionary representation of a BaseModel.
         """
+        # mypy doesn't realize that BaseModel has a field called __dataclass_fields__
+        dataclass_fields = self.__dataclass_fields__.keys()  # type: ignore
         return {
-            key: self._to_dict_elem(value)
-            for key, value in self.__dict__.items()
-            if value is not None
+            key: self._to_dict_elem(getattr(self, key))
+            for key in dataclass_fields
+            if getattr(self, key) is not None
         }
 
     def _to_dict_elem(self: BaseModel, elem: Any) -> Any: